tt-adapter/pyproject.toml
[project]
name = "tt-adapter"
version = "0.0.1"

dependencies = [
  "ai-edge-model-explorer"
]